test_server improvements
When running in UNIT_TEST mode, only run the tests of the tested addon, not of
its dependencies. This saves time as the tests of the dependencies are not run
again and again for each addon being tested. We keep the current behavior in
non UNIT_TEST mode to let Odoo sort out the modules dependencies (this is
important for a repository containing e.g. both sale and sale_stock: if both
addons are installed first, and then tested, the tests of sale will fail. The
case does not happen in UNIT_TEST model)

Use the list syntax of subprocess rather than interpolated strings

Use subprocess.Popen rather than a subshell to handle the pipe.