[ADD] Add new test named test_odoo_lint that run test over the addons modules and check the next conditions:
    - Module description (legacy mode, rst syntax, exist README.rst).
    - Module icon exist.
    - Module doc exist and check index.rst syntax.

The test_odoo_lint script return a log that check all the conditions per
module and show instant the error found. At the end of the script will show
an error summary and a warning summary indicating the list of module that
have errors. This error summary is split by type:

    - Description README.rst syntax.
    - Missing README.rst file
    - Missing icon.png as module icon.
    - Documentation folder but not index.rst file.
    - Documentation index.rst syntax.

Also a new enviroment variable ODOO_LINT was added as part of the
travis_run_tests to run test_odoo_lint.