allow to ignore a file
* allow to ignore a file

Pylint allows you to ignore a file completely adding a "# pylint: disable-all"
line at the top. This is especially useful in cases where a file is
copied for some reason and we don't want to mess too much with it to
avoid diverging from the original. This option make the "disable-all"
option an error, rather confusingly.