[REF] include recursive addons and module search
#392 introduced a two level search on addons, but ommitted to adapt some uses of get_modules
who where expecting addon folders with modules at it's first level

This commit tries to solve the problem in a more abstractable manner and introduces recursive
searches for modules and addons. As the addons search depends on the detection of modules in
its direct subdirectories aligning both searches in test_server assures consistent module and
addons detection.

This commit stops at the first level it detects an addon or a module, so that we can have multiple
folder structure, but in the end all addons and modules must be on it's respective uniq level in
order to be detected all.