[REF] install: Install requests[security] again to avoid new errors (#457)
* [REF] install: Install requests[security] to avoid new errors

Travis use a old python2.7 without security patches to httplib built-in package.
Then show errors "error: [Errno 111] Connection refused" for get urls

A upgrade of Python2.7.LATEST could fix this issue but we can not do it
for travis.
A requests[security] installation should fix this error, so is easiest.

* fixup! [REF] install: Install requests[security] to avoid new errors

* [REF] clone_oca_dependencies: Use --upgrade parameter to install dependencies with custom version