[IMP] test_server: Patch odoo to create unlogged DB tables
This commit causes Odoo to be patched so all database tables are created
as unlogged.

This is applied only for testing instances because, if a docker
container is stopped without firstly saving the changes, all records
will be deleted; so it's not suitable for working under other
environments, e.g. locally.

This is compatible with the odoo versions >=7.0.

Co-authored-by: Luis González <lgonzalez@vauxoo.com>