[MERGE] Merge from branch OCA/mqt:master (#4)
* [REF] requirements: Supports code-block for module/README.rs
Closes https://github.com/OCA/maintainer-quality-tools/issues/568

* [FIX] self_tests: Adjust expected values after pylint change

In a recent change on pylint-odoo [1], it was introduced a fix to
recognize a badly-formatted version in manifest, which now depends on
the Odoo version. When the version is provided and it's a numeric
version value (e.g. not "master") it's taken into account to determine
whether the version defined in the manifest is formatted correctly.

Since versions in tested modules are as follows:
- Two set to 8.0
- Four set to wrong values (e.g. 1.0)

When the version is numeric but it's not 8.0, there will be six errors,
otherwise there will be four. In these tests, the only case when it's
numeric but it's not 8.0, is when testing 7.0.

[1] Pylint change: https://github.com/OCA/pylint-odoo/commit/6b1f1e1bfaf

* [REF] cfg: Rename param manifest_required_author

Since it was renamed on OCA/pylint-odoo@8dcb1eb9df1e, from
`manifest_required_author` to `manifest_required_authors` (from singular
to plural), it needs to be renamed here, too.

* [FIX] Make happy manifest_required_authors syntax

We shouldn't have quotes

* [REF] .travis.yml: Use python 3.7

https://docs.travis-ci.com/user/languages/python/#Specifying-Python-versions

* [IMP] travis_install_nightly: support wkhtmltopdf 0.12.5

Currently, only those versions of wkhtmltopdf that come on a .tar.xz are
installable, but not those who come on .deb packages. That includes the
most recent version 0.12.5.

This commit adds support for those versions, which includes 0.12.1 and
0.12.5

Closes #572

* [FIX] test_server: Ignore trailing whitespaces on log messages

Odoo recently introduced a change in the way logs are printed, adding
performance information when logging HTTP requests:
https://github.com/odoo/odoo/commit/c44fe912329e2d4c7541ed810526d4a069f930c4#diff-d301b3dd399c26a7f38fb31c9e1dfbddR136

However, that causes all messages who don't containg such information to
contain a trailing whitespace.

This commit improves the regexp used to match messages in logs, to not
take into account trailing whitespaces.#

* [IMP] test_server: Improve error msg when "Modules loaded." is not found

This improves the error message, so it's clearer.

* [REF] requirements: Supports new frontend unittest engine (#578)

* [REF] requirements: Supports new frontend unittest engine

* Fix the following warning:
  - websocket-client module is not installed

* Add a new 12.0 environment

* Enable makepot on 12.0 branch

* remove whitespace before ( for E211

* [FIX] Add six dependency

* [FIX] never break for decoding errors in log

* [REF] pylint-cfg: enable redefined-builtin pylint check
Closes https://github.com/OCA/pylint-odoo/issues/238

* [FIX] Call the constructor with proper params for being compatible with all versions