[IMP] clone_oca_dependencies: Clone repositories from commits SHA
This improvement allows to clone repositories selecting a commit SHA.
Useful when you want to work with an specific state of the repository
that is not marked by any branch.

The syntax of the line you have to place in oca_dependencies.txt is as
follows:

```
server-tools git@github.com:OCA/server-tools.git 12.0 0c2c386

```

Note that you need to explicitly indicate a branch where the commit is
present because git doesn't allows to clone or fetch using directly a
commit SHA without having a configuration parameter set at server side[1].

So, when you choose to clone a repository using a commit SHA,
`clone_oca_dependencies` will fetch the commit history for the branch
you selected and then reset `HEAD` position to that commit.

[1] https://stackoverflow.com/questions/31278902/how-to-shallow-clone-a-specific-commit-with-depth-1/43136160#43136160