[IMP] .coveragerrc: Allow to control decimal precision & min through env (#314)
* [IMP] .coveragerrc: Allow to control decimal precision & min through env

It's now possible to control the following using environment variables:
- Decimal precision, usin $COVERAGE_PRECISION
- Minimum allowed coverage, using $COVERAGE_MIN

When the above variables are not set , values will fallback to their
defaults, i.e. 0.

For more information, please check the configuration reference of
Coverage.py:
https://coverage.readthedocs.io/en/coverage-5.0/config.html

* [IMP] .coveragerrc: Allow to control decimal precision & min through env

It's now possible to control the following using environment variables:
- Decimal precision, usin $COVERAGE_PRECISION
- Minimum allowed coverage, using $COVERAGE_MIN
- Custom paths to exclude, using $EXCLUDE_COVERAGE

When the above variables are not set , values will fallback to their
defaults, i.e. 0 in case of numeric values, or an empty string in case
of custom paths (which has no effect).